pyproject.toml
[tool.poetry]
name = "django_jwt"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.11.4"
django = "^4.1.6"
djangorestframework = "^3.14.0"
PyJWT = "^2.6.0"
psycopg2 = "^2.9.5"
redis = "^4.5.5"
django-redis = "^5.3.0"
psycopg2-binary = "^2.9.6"

